% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{compare1,matchdesign-method}
\alias{compare1,matchdesign-method}
\title{compare1 method for the matchdesign class}
\usage{
\S4method{compare1}{matchdesign}(x, ...)
}
\description{
Measures for comparison of empirical distributions of common variables in an univariate sense
}

