% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{compare2,matchdesign-method}
\alias{compare2,matchdesign-method}
\title{compare2 method for the matchdesign class}
\usage{
\S4method{compare2}{matchdesign}(x, ...)
}
\description{
Measures for comparison of empirical distributions of common variables in an univariate sense, by strata
}

