% Generated by roxygen2 (4.0.2): do not edit by hand
\name{compare_matchvars}
\alias{compare_matchvars}
\title{Compare matching variables between two \code{filetomatch} objects}
\usage{
compare_matchvars(x = "filetomatch", y = "filetomatch", ...)
}
\arguments{
\item{x}{\code{filetomatch} object}

\item{y}{\code{filetomatch} object}

\item{type}{type of desired result: table, plot or disimilarity measures}

\item{weights}{logical (\code{TRUE} or \code{FALSE}) to indicate if weights are to be used in the comparison}

\item{strata}{logical (\code{TRUE} or \code{FALSE}) to indicate if comparison is to be made by levels of a strata variable}
}
\value{
Tables, plots or measures.
}
\description{
\code{compare_matchvars} makes a comparison of the empirical
distributions observed in two separate \code{\link{filetomatch}} objects for the
list of defined \code{matchvars}.
}
\details{
Using weights is optional. Also, the comparison can be made by levels of
a strata variable.

Note that the strata variable must be defined in the \code{stratavars} slots in
the \code{\link{filetomatch}} variables.
}
\seealso{
Other "Select matching variables": \code{\link{compare_var_strata}};
  \code{\link{compare_var}}; \code{\link{plot2cat}};
  \code{\link{predictvalue_var}};
  \code{\link{predictvalue}}; \code{\link{similarity2cat}};
  \code{\link{tabulate2cat}}; \code{\link{uncert2vars}}
}

