% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{compare_var,filetomatch,filetomatch-method}
\alias{compare_var,filetomatch,filetomatch-method}
\title{Compare a (single) shared variable across two 'filetomatch' objects}
\usage{
\S4method{compare_var}{filetomatch,filetomatch}(x, y, var_A, var_B,
  type = "table", weights = FALSE, ...)
}
\description{
Compare a (single) shared variable across two 'filetomatch' objects
}
\seealso{
Other 'Select matching variables': \code{\link{compare_matchvars,filetomatch,filetomatch-method}};
  \code{\link{compare_var_strata,filetomatch,filetomatch-method}};
  \code{\link{plot2cat}};
  \code{\link{predictvalue,filetomatch-method}};
  \code{\link{similarity2cat}}
}

