% Generated by roxygen2 (4.0.2): do not edit by hand
\name{compare_var}
\alias{compare_var}
\title{Generic: "compare_var"
applies to pairs of 'filetomatch' objects}
\usage{
compare_var(x = "filetomatch", y = "filetomatch", ...)
}
\description{
Generic: "compare_var"
applies to pairs of 'filetomatch' objects
}

