% Generated by roxygen2 (4.0.2): do not edit by hand
\name{compare_var_strata}
\alias{compare_var_strata}
\title{Compare a (single) shared variable across two \code{filetomatch} objects
by groups defined by the strata variable}
\usage{
compare_var_strata(x = "filetomatch", y = "filetomatch", ...)
}
\arguments{
\item{x}{\code{filetomatch} object}

\item{y}{\code{filetomatch} object}

\item{var_A}{name of variable in the first file}

\item{var_B}{name of variable in the second file}

\item{type}{type of desired result: table, plot or disimilarity measures}

\item{weights}{logical (\code{TRUE} or \code{FALSE}) to indicate if weights are to be used in the comparison}

\item{stratavar}{name of the strata variable.}
}
\value{
Table, plot or measures by strata groups
}
\description{
\code{compare_var_strata} function compares a single variable
across two separate \code{\link{filetomatch}} objects by considering groups
defined by the \code{stratavars} variable in the \code{filetomatch} objects.
}
\details{
The name of the variables in each of the files (\code{var_A} and \code{var_B})
could be different, but the categories must be exactly the same in order to
avoid problems when making the comparison.

Note that both \code{stratavars} and \code{weights} names (if used)
must be equal in the two \code{filetomatch} objects.
}
\seealso{
Other "Select matching variables": \code{\link{compare_matchvars}};
  \code{\link{compare_var}}; \code{\link{plot2cat}};
  \code{\link{predictvalue_var}};
  \code{\link{predictvalue}}; \code{\link{similarity2cat}};
  \code{\link{tabulate2cat}}; \code{\link{uncert2vars}}
}

