% Generated by roxygen2 (4.0.1): do not edit by hand
\name{compareMultivar}
\alias{compareMultivar}
\title{Comparing multivariate distributions of categorical variables in two files}
\usage{
compareMultivar(var1A, var1B, var2A, var2B, var3A = NULL, var3B = NULL,
  fileA, fileB, wA = NULL, wB = NULL, plot = FALSE, type = "abs",
  measures = FALSE, ...)
}
\arguments{
\item{var1A}{name of first variable (independet) in the first file}

\item{var1B}{name of first variable (independent) in the second file}

\item{var2A}{name of second variable (dependent) in the first file}

\item{var2B}{name of second variable (dependent) in the first file}

\item{var3B}{name of third variable (independent) in the first file (optional)}

\item{var3B}{name of third variable (independet) in the first file (optional)}

\item{fileA}{data frame containing first file}

\item{fileB}{data frame containing second file}

\item{wA}{name of variable containing weights for the first file (optional)}

\item{wB}{name of variable containing weights for the second file (optional)}

\item{plot}{logical value. TRUE if a plot is desired, FALSE otherwise. Defaults to FALSE.}

\item{type}{character value if x1, x2 character vectors: type of values desired for cells ('abs' or 'rel').}

\item{measures}{similarity measures for categorical x1, x2 or descriptives for numerical x1, x2}
}
\value{
list: table, measures, plot (if plot=TRUE)
}
\description{
Compares distribution categorical variables (maximum 3) in two distinct files
}
\details{
Dependencies: StatMatch, Hmisc, ggplot2, vcd
}
\seealso{
Other "Select variables": \code{\link{plot2Cat}};
  \code{\link{plotTable}}; \code{\link{predictvalue}};
  \code{\link{uncertvarxvary}}
}

