% Generated by roxygen2 (4.0.1): do not edit by hand
\name{compareVar}
\alias{compareVar}
\title{Comparing distributions of a variable in distinct microdata files}
\usage{
compareVar(varA, varB, fileA, fileB, wA = NULL, wB = NULL, plot = FALSE,
  compareas = "categorical", type = "abs", measures = FALSE, ...)
}
\arguments{
\item{varA}{name of the variable in the first file}

\item{varB}{name of the variable in the second file}

\item{wA}{name of variable containing weights for the first file (optional)}

\item{wB}{name of variable containing weights for the second file (optional)}

\item{compareas}{Indicates how variables are to be compared: 'categorical' or 'numeric'}

\item{plot}{logical value. TRUE if a plot is desired, FALSE otherwise. Defaults to FALSE.}

\item{type}{character value if x1, x2 character vectors: type of values desired for cells ('abs' or 'rel').}

\item{measures}{similarity measures for categorical x1, x2 or descriptives for numerical x1, x2}
}
\value{
list: table, measures, plot (if plot=TRUE)
}
\description{
This function sompares the observed distribution of a variable (continuous or categorical)
in two distinct, independent microdata files
}
\details{
Put more details, if necessary
}
\author{
"Ines Garmendia <ines.garmendia@gmail.com>"
}

