% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{concatenate,filetomatch,filetomatch-method}
\alias{concatenate,filetomatch,filetomatch-method}
\title{Concatenate two 'filetomatch' objects}
\usage{
\S4method{concatenate}{filetomatch,filetomatch}(x, y)
}
\description{
Concatenate two 'filetomatch' objects
}
\seealso{
Other 'Apply matching method': \code{\link{match.hotdeck,filetomatch,filetomatch-method}}
}

