% Generated by roxygen2 (4.0.2): do not edit by hand
\name{concatenate}
\alias{concatenate}
\title{Concatenate two \code{filetomatch} objects}
\usage{
concatenate(x = "filetomatch", y = "filetomatch", ...)
}
\arguments{
\item{x}{\code{filetomatch} object}

\item{y}{\code{filetomatch} object}
}
\value{
A \code{\link{fusedfile}} object for which \code{data} that contains all
rows in both \code{data} slots, and in which non-observed values are filled with \code{NA}'s.
}
\description{
\code{concatenate} takes data from two \code{\link{filetomatch}} objects
and produces a \code{\link{fusedfile}} object which contains data from both sources.
}
\details{
In the resulting \code{\link{fusedfile}} object, \code{data} contains
all rows (observations) and columns (variables) in the original files, as well as
additional information to trace back the origin of variables.

Specifically:
\itemize{
     \item \code{origin_specvars} stores the source of \code{specvars} in the original files, in corresponding order.
     \item \code{origin_weights} stores the source of \code{weights} in the original files, in corresponding order.
}
}
\seealso{
Other "Apply matching method": \code{\link{fusedfile-class}};
  \code{\link{match.hotdeck}}
}

