% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{describe,matchdesign-method}
\alias{describe,matchdesign-method}
\title{Summary method for the matchdesign class}
\usage{
\S4method{describe}{matchdesign}(x, ...)
}
\description{
Method that gives the number of rows, commmon (matching) and specific variables
}

