% Generated by roxygen2 (4.0.2): do not edit by hand
\name{donor}
\alias{donor}
\title{Create donor file object}
\usage{
donor(data, matchvars, specvars, stratavars, weights)
}
\arguments{
\item{data}{data frame where rows are observations and columns contain common & specific variables, and possibly weights}

\item{matchvars}{character vector with names of common (matching) variables}

\item{specvars}{character vector with names of specific variables}

\item{stratavars}{character vector with names of variables acting as strata (optional)}

\item{weights}{character vector with the name of the vector with numeric weights (optional, but usually given in survey samples)}
}
\value{
Object of class filetomatch with donor role
}
\description{
Creates a donor file according to 'filetomatch' class definition
}
\details{
This function creates an object of class 'filetomatch' for a microdata file
acting as the donor file in matching. It this thus assumed that a receptor file will be created
at the same time.

All the necessary information for matching should be provided:
data, identification of common & specific variables (compulsory), identification of a weight variable (optional), and identification
of strata variables (optional).
}
\author{
"Ines Garmendia <ines.garmendia@gmail.com>"
}
\seealso{
Other "Matching definition step": \code{\link{filetomatch-class}};
  \code{\link{receptor}}; \code{\link{symmetric}}
}

