% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{ecv}
\alias{ecv}
\title{Living Conditions Survey data}
\format{data frame with 4749 rows (observations) and 26 columns (variables).}
\source{
\href{http://en.eustat.es/estadisticas/tema_172/opt_0/ti_Condiciones_de_vida/temas.html#axzz37iVNbE00}{Eustat}
}
\usage{
data(ecv)
}
\description{
Microdata file containing data from the Living Conditions Survey
(Encuesta de Condiciones de Vida, ECV), as provided by Eustat - the Basque Statistical Office (Spain)

The reference period is the 4th quarter of 2009,
and the reference population is people living in households at the Basque Autonomous Community,
}
\details{
The original microdata contains 5242 observations, which are reduced to 4749 after
filtering people aged 16 and over. A set of 26 variables was selected for inclusion in this package: common variables
with pra data, a weight variable and 13 items measuring some aspect of living conditions.

Common variables

\code{territory} Territory: Araba, Gipuzkoa, Bizkaia.

\code{age} Age (6 categories)

\code{sex} Sex

\code{agesex} Combination of age and sex (12 categories)

\code{famsize} Family size: 1, 2, 3 or more

\code{student} Student: yes or no? (logical vector)

\code{seekjob} Seeking for a job: yes or no? (logical vector)

\code{employed} Employed: yes or no? (logical vector)

\code{unemployed} Unemployed: yes or no? (logical vector)

\code{inactive} Inactive: yes or no? (logical vector)

\code{workhours} Working hours (numerical). Only for occupied.

\code{housework} Dedication to housework.
}
\author{
"Ines Garmendia <ines.garmendia@gmail.com>"
}
\keyword{datasets}

