% Generated by roxygen2 (4.0.2): do not edit by hand
\name{frechet.uncertainty}
\alias{frechet.uncertainty}
\title{Frechet uncertainty bounds for categorical variables observed in separate files}
\usage{
frechet.uncertainty(x = "filetomatch", y = "filetomatch", ...)
}
\arguments{
\item{x}{a \code{\link{filetomatch}} object}

\item{y}{a \code{\link{filetomatch}} object}

\item{var_x}{Specific variable in \code{x}}

\item{var_y}{Specific variable in \code{y}}

\item{print.f}{type desired result: "tables" (default) or "data.frame"}

\item{tol}{tolerance used in comparing joint distributions for common (Z) variables, stored as \code{matchvars}}
}
\description{
here,
}

