% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{fusedfile-class}
\alias{fusedfile-class}
\title{Microdata file after matching or concatenation w.r.t. another microdata file}
\description{
Microdata file after matching or concatenation w.r.t. another microdata file
}
\seealso{
Other "Matching definition step": \code{\link{filetomatch-class}};
  \code{\link{genericmatch-class}}
}

