% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{genericmatch-class}
\alias{genericmatch-class}
\title{Generic matching class}
\description{
Generic matching class
}
\seealso{
Other "Matching definition step": \code{\link{filetomatch-class}};
  \code{\link{fusedfile-class}}
}

