% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{include,filetomatch-method}
\alias{include,filetomatch-method}
\title{Include variables to a 'filetomatch' object}
\usage{
\S4method{include}{filetomatch}(x, vars, as, ...)
}
\description{
Include variables to a 'filetomatch' object
}
\seealso{
Other 'Utilities': \code{\link{remove,filetomatch-method}};
  \code{\link{select_observations,filetomatch-method}};
  \code{\link{select_strata,filetomatch-method}}
}

