% Generated by roxygen2 (4.0.2): do not edit by hand
\name{include}
\alias{include}
\title{Add variables to a \code{filetomatch} object}
\usage{
include(x = "filetomatch", ...)
}
\arguments{
\item{x}{a \code{\link{filetomatch}} object}

\item{vars}{names of variables to be removed (can be more than one)}

\item{as}{name of slot where variables are to be added: \code{matchvars}, \code{specvars}, \code{stratavars} or \code{weightvars}}
}
\description{
\code{include} creates a new \code{\link{filetomatch}} object by adding the
indicated variables to a \code{filematch} object.

The type of varible has to be specified by means of the \code{as} parameter value:
\code{matchvars}, \code{stratavars}, or \code{weightvars}.
}
\seealso{
Other "Utilities": \code{\link{remove}};
  \code{\link{select_observations}};
  \code{\link{select_strata}}; \code{\link{update}}
}

