% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{match.hotdeck,filetomatch,filetomatch-method}
\alias{match.hotdeck,filetomatch,filetomatch-method}
\title{Match two 'filetomatch' objects via distance hot-deck}
\usage{
\S4method{match.hotdeck}{filetomatch,filetomatch}(x, y)
}
\description{
Match two 'filetomatch' objects via distance hot-deck
}
\seealso{
Other 'Apply matching method': \code{\link{concatenate,filetomatch,filetomatch-method}}
}

