% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{matchdesign-class}
\alias{matchdesign}
\alias{matchdesign-class}
\title{Specify matching design object}
\arguments{
\item{rec}{microdata file acting as receptor file A}

\item{don}{microdata file acting as donor file B}

\item{matchvars}{list of common matching variables}

\item{recvars}{list of specific variables selected in receptor file A}

\item{donvars}{list of specific variables selected in donor file B}
}
\description{
Specify a new matching project to link specific variables from two microdata files based on common variables between them
}

