% Generated by roxygen2 (4.0.1): do not edit by hand
\name{nnhdbystrata}
\alias{nnhdbystrata}
\title{Create fused file using NN hot deck for a given strata and a given list of categorical matching variables}
\usage{
nnhdbystrata(rec, don, stratalevel, stratavar, matchvars, vary,
  checkdiffs = FALSE)
}
\arguments{
\item{rec}{recipient file data frame}

\item{don}{donor file data frame}

\item{stratalevel}{character identifying strata: a level of stratavar}

\item{stratavar}{name of the variable to define the strata given by strataid}

\item{matchvars}{character vector with names of matching variables}

\item{vary}{variables in donor file to be imputed to the recipient file}

\item{checdiffs}{logical value indicating if a check of differences between donor-recipient files is desired}
}
\value{
Fused file
}
\description{
For a given strata and a list of variables, it performs nearest neighbour hot deck matching for categorical variables using Gower's distance and constrained matchinig
}
\author{
"Ines Garmendia"
}

