% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plot2cat}
\alias{plot2cat}
\title{Compare a single categorical variable across two data frames: Plot}
\usage{
plot2cat(data_A, data_B, var_A, var_B, weights_A = NULL, weights_B = NULL,
  cell_values = "abs")
}
\description{
Barplots for a single categorical variable in two distinct, independent data frames
}
\seealso{
Other "Select matching variables": \code{\link{compare_matchvars}};
  \code{\link{compare_var_strata}};
  \code{\link{compare_var}};
  \code{\link{predictvalue_var}};
  \code{\link{predictvalue}}; \code{\link{similarity2cat}};
  \code{\link{tabulate2cat}}; \code{\link{uncert2vars}}
}

