% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plotCat}
\alias{plotCat}
\title{Barplot for a categorical variable}
\usage{
plotCat(data, type = "rel")
}
\arguments{
\item{data}{data frame result of as.data.frame(xtabs()). 1st col x: levels of variable; 2nd col y: counts or freqs}

\item{type}{type of values associated with the levels of the factor: absolute (abs) or relative (rel)}
}
\value{
a graphical object (barplot)
}
\description{
Internal function that computes barplot for a factor variable,
to be used inside other functions in micromatch.
}
\details{
Details here if necessary.
}
\author{
"Ines Garmendia <ines.garmendia@gmail.com>"
}
\seealso{
Other "Select variables": \code{\link{compareMultivar}};
  \code{\link{plotTable}}; \code{\link{predictvalue}};
  \code{\link{uncertvarxvary}}
}
\keyword{graphics}

