% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plotTable}
\alias{plotTable}
\title{Visual representation of a 2-way&3-way contingency tables based on vcd structplot}
\usage{
plotTable(table)
}
\arguments{
\item{table}{xtabs object crossing the three variables (usually the specific will be the last), cells can be absolute or relative values.}
}
\value{
a graphical object (structplot)
}
\description{
Internal function to visualize 3-way contingency tables, to be used inside other functions in micromatch. Usage: two ways are for strata variables (e.g. sex and age) and third for variable of interest (common or specific)
}
\author{
"Ines Garmendia <ines.garmendia@gmail.com>"
}
\seealso{
Other "Select variables": \code{\link{compareMultivar}};
  \code{\link{plot2Cat}}; \code{\link{predictvalue}};
  \code{\link{uncertvarxvary}}
}
\keyword{graphics}

