% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{pra}
\alias{pra}
\title{Population with Relation to Activity Survey data}
\format{data frame with 10865 rows (observations) and 73 columns (variables).}
\source{
microdata for the Population in Relation to Activity survey, period 2009, 4th quarter. Reference page:
\link{http://en.eustat.es/ci_ci/estadisticas/tema_37/opt_0/tipo_1/ti_Population_with_relation_to_activity_PRA/temas.html#axzz37idlbZx5}
}
\description{
Microdata file containing data from the Population in Relation to Activity survey panel
(Encuesta de Población en Relación con la Actividad, PRA),
as provided by Eustat - the Basque Statistical Office \link{http://en.eustat.es/ci_ci/indice.htlm}.

The reference period is the 4th quarter of 2009,
and the reference population is people living in households at the Basque Autonomous Community (Spain).
}
\details{
The original microdata contains 12658 observations, which are reduced to 10865 after
filtering people aged 16 and over. Explain main variables here.
}
\author{
Ines Garmendia.
}

