% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{predictvalue,filetomatch-method}
\alias{predictvalue,filetomatch-method}
\title{Assess predictive value of common variables w.r.t specific variables in a 'filetomatch' object}
\usage{
\S4method{predictvalue}{filetomatch}(x = "filetomatch", ...)
}
\description{
Assess predictive value of common variables w.r.t specific variables in a 'filetomatch' object
}
\seealso{
Other 'Select matching variables': \code{\link{compare_matchvars,filetomatch,filetomatch-method}};
  \code{\link{compare_var,filetomatch,filetomatch-method}};
  \code{\link{compare_var_strata,filetomatch,filetomatch-method}};
  \code{\link{plot2cat}}; \code{\link{similarity2cat}}
}

