% Generated by roxygen2 (4.0.1): do not edit by hand
\name{predictvalue}
\alias{predictvalue}
\title{Evaluating predictive value of individual variables w.r.t. a specific variable}
\usage{
predictvalue(varx, vary, data, varw = NULL, plot = FALSE,
  measures = FALSE, type = "abs", ...)
}
\arguments{
\item{varx}{name of common variable}

\item{vary}{name of specific variable}

\item{data}{name of data-frame containing both x and y}

\item{varw}{name of weights variable in data-frame data (optional)}

\item{plot}{logical value. TRUE if a plot is desired, FALSE otherwise. Defaults to FALSE.}

\item{measures}{logical value. TRUE if measures of predictive value are desired.}

\item{type}{type of cell values desired in the table: 'abs' or 'rel'}
}
\value{
list: table, measures, plot
}
\description{
Evaluates predictive value of a common variable \code{x} with respect to a specific variable \code{y}, in the context of statistically matching two independent files.
}
\details{
Dependencies: StatMatch, Hmisc, ggplot2
}
\seealso{
Other "Select variables": \code{\link{compareMultivar}};
  \code{\link{plot2Cat}}; \code{\link{plotTable}};
  \code{\link{uncertvarxvary}}
}
