% Generated by roxygen2 (4.0.2): do not edit by hand
\name{predictvalue_var}
\alias{predictvalue_var}
\title{Assess predictive value of a variable w.r.t. another variable in a data frame}
\usage{
predictvalue_var(data, vars_x, var_y, weights = NULL)
}
\arguments{
\item{data}{name of data frame containing both x and y}

\item{vars_x}{name of common variable(s)}

\item{var_y}{name of specific variable}

\item{weights}{weight vector name (optional)}
}
\description{
Evaluates predictive value of a common variable \texttt{x} with respect to a specific variable \texttt{y}
}
\details{
Dependencies: StatMatch
}
\seealso{
Other "Select matching variables": \code{\link{tabulate2cat}}
}

