% Generated by roxygen2 (4.0.2): do not edit by hand
\name{receptor}
\alias{receptor}
\title{Create receptor file object}
\usage{
receptor(data, matchvars, specvars, stratavars, weights)
}
\arguments{
\item{data}{data frame}

\item{matchvars}{character vector with names of common (matching) variables}

\item{specvars}{character vector with names of specific variables}

\item{stratavars}{character vector with names of variables acting as strata (optional)}

\item{weights}{character vector with the name of the vector with numeric weight (optional, but usually given in survey samples)}
}
\value{
Object of class filetomatch with receptor role
}
\description{
Creates a receptor file according to 'filetomatch' class definition.
}
\details{
This function creates an object of class 'filetomatch' for a microdata file
acting as the receptor file in matching. It is thus assumed that a donor file will be created a the same time.

All the necessary information for matching should be provided:
data, identification of common & specific variables (compulsory), identification of a weight variable (optional), and identification
of strata variables (optional).
}
\author{
"Ines Garmendia <ines.garmendia@gmail.com>"
}
\seealso{
Other "Matching definition step": \code{\link{donor}};
  \code{\link{filetomatch-class}}; \code{\link{symmetric}}
}

