% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{remove,filetomatch-method}
\alias{remove,filetomatch-method}
\title{Remove variables from a 'filetomatch' object}
\usage{
\S4method{remove}{filetomatch}(x, vars, ...)
}
\description{
Remove variables from a 'filetomatch' object
}
\seealso{
Other 'Utilities': \code{\link{include,filetomatch-method}};
  \code{\link{select_observations,filetomatch-method}};
  \code{\link{select_strata,filetomatch-method}}
}

