% Generated by roxygen2 (4.0.2): do not edit by hand
\name{remove}
\alias{remove}
\title{Remove variables from a \code{filetomatch} object}
\usage{
remove(x = "filetomatch", ...)
}
\arguments{
\item{x}{a \code{\link{filetomatch}} object}

\item{vars}{names of variables to be removed (can be more than one)}
}
\description{
\code{remove} creates a new \code{\link{filetomatch}} object by removing the
indicated variables from the corresponding place in the \code{filematch} objects:
\code{matchvars}, \code{stratavars}, or \code{weightvars}.

Note that the new object must be assigned to a name in order to re-use it.
}
\seealso{
Other "Utilities": \code{\link{include}};
  \code{\link{select_observations}};
  \code{\link{select_strata}}; \code{\link{update}}
}

