% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{select_observations,filetomatch-method}
\alias{select_observations,filetomatch-method}
\title{Select observations (i.e rows of data frame) in a 'filematch' object}
\usage{
\S4method{select_observations}{filetomatch}(x, obs, ...)
}
\description{
Select observations (i.e rows of data frame) in a 'filematch' object
}
\seealso{
Other 'Utilities': \code{\link{include,filetomatch-method}};
  \code{\link{remove,filetomatch-method}};
  \code{\link{select_strata,filetomatch-method}}
}

