% Generated by roxygen2 (4.0.2): do not edit by hand
\name{select_observations}
\alias{select_observations}
\title{Select data for a strata value in a \code{filetomatch} object}
\usage{
select_observations(x = "filetomatch", ...)
}
\arguments{
\item{x}{a \code{\link{filetomatch}} object}

\item{obs}{indices of rows to keep in \code{data}}
}
\description{
\code{select_strata} creates a new \code{\link{filetomatch}} object
by selecting the indicated rows value for the \code{data}.

The result is a new \code{filematch} object where rows have been filtered for the
as indicated.
}
\seealso{
Other "Utilities": \code{\link{include}};
  \code{\link{remove}}; \code{\link{select_strata}};
  \code{\link{update}}
}

