% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{select_strata,filetomatch-method}
\alias{select_strata,filetomatch-method}
\title{Select strata values in a 'filematch' object}
\usage{
\S4method{select_strata}{filetomatch}(x, value, ...)
}
\description{
Select strata values in a 'filematch' object
}
\seealso{
Other 'Utilities': \code{\link{include,filetomatch-method}};
  \code{\link{remove,filetomatch-method}};
  \code{\link{select_observations,filetomatch-method}}
}

