% Generated by roxygen2 (4.0.2): do not edit by hand
\name{select_strata}
\alias{select_strata}
\title{Select data for a strata value in a \code{filetomatch} object}
\usage{
select_strata(x = "filetomatch", ...)
}
\arguments{
\item{x}{a \code{\link{filetomatch}} object}

\item{value}{wanted value to filter with \code{stratavars} variable}
}
\description{
\code{select_strata} creates a new \code{\link{filetomatch}} object
by selecting the indicated strata value for a \code{stratavars} variable
in the object.

The result is a new \code{filematch} object where rows have been filtered for the
by the indicated strata value in \code{data}.
}
\seealso{
Other "Utilities": \code{\link{include}};
  \code{\link{remove}}; \code{\link{select_observations}};
  \code{\link{update}}
}

