% Generated by roxygen2 (4.0.2): do not edit by hand
\name{similarity2cat}
\alias{similarity2cat}
\title{Compare a single categorical variable across two data frames: Similarity/Disimilarity measures}
\usage{
similarity2cat(data_A, data_B, var_A, var_B, weights_A = NULL,
  weights_B = NULL)
}
\description{
Similarity/disimilarity measures based on observed empirical distributions in the two data frames.
}
\details{
Details about parameter values for Statmatch comp.prop function
}
\seealso{
Other 'Select matching variables': \code{\link{compare_matchvars,filetomatch,filetomatch-method}};
  \code{\link{compare_var,filetomatch,filetomatch-method}};
  \code{\link{compare_var_strata,filetomatch,filetomatch-method}};
  \code{\link{plot2cat}};
  \code{\link{predictvalue,filetomatch-method}}
}

