% Generated by roxygen2 (4.0.2): do not edit by hand
\name{symmetric}
\alias{symmetric}
\title{Create matching file object with no specific (receptor nor donor) role}
\usage{
symmetric(data, matchvars, specvars, stratavars, weights)
}
\arguments{
\item{data}{data frame where rows are observations and columns contain common & specific variables, and possibly weights}

\item{matchvars}{character vector with names of common (matching) variables}

\item{specvars}{character vector with names of specific variables in data}

\item{stratavars}{character vector with names of variables acting as strata (optional)}

\item{weights}{character vector with the name of the vector with numeric weights (optional, but usually given in survey samples)}
}
\value{
Object of class filetomatch with symetric role
}
\description{
Creates a matching file object with no specific role, according to 'filetomatch' class definition
}
\details{
This function creates an object of class 'filetomatch' for a microdata file
to be matched to another, independent file. The file is supposed to lack a specific role in matching,
namely (as in the case of receptor and donor pairs), that is, the aim in this case is to fill all the rows in both files.

All the necessary information for matching should be provided:
data, identification of common & specific variables (compulsory),
identification of weight variable (optional), and identification
of strata variables (optional).
}
\author{
"Ines Garmendia <ines.garmendia@gmail.com>"
}
\seealso{
Other "Matching definition step": \code{\link{donor}};
  \code{\link{filetomatch-class}}; \code{\link{receptor}}
}

