% Generated by roxygen2 (4.0.2): do not edit by hand
\name{tabulate2cat}
\alias{tabulate2cat}
\title{Compare a single categorical variable across two data frames: Tabulate}
\usage{
tabulate2cat(data_A, data_B, var_A, var_B, weights_A = NULL,
  weights_B = NULL, cell_values = "abs")
}
\arguments{
\item{data_A}{first data frame}

\item{data_B}{second, distinct, data frame}

\item{var_A}{name of the variable in the first file, data_A}

\item{var_B}{name of the variable in the second file, data_B}

\item{weights_A}{name of variable containing weights for the first file, data_A (optional)}

\item{weights_B}{name of variable containing weights for the second file, data_B (optional)}

\item{cell_values}{type of values desired for cells: 'abs' (absolute values) or 'rel' (relative values).}
}
\value{
table
}
\description{
Frequency tables for a categorical variable in two distinct, independent data frames
}
\details{
Details.
}
\author{
Ines Garmendia <ines.garmendia@gmail.com>
}
\seealso{
Other "Select matching variables": \code{\link{compare_matchvars}};
  \code{\link{compare_var_strata}};
  \code{\link{compare_var}}; \code{\link{plot2cat}};
  \code{\link{predictvalue_var}};
  \code{\link{predictvalue}}; \code{\link{similarity2cat}};
  \code{\link{uncert2vars}}
}

