% Generated by roxygen2 (4.0.1): do not edit by hand
\name{tabvisvar}
\alias{tabvisvar}
\title{Tabulate and visualize ECV(LCS) specific variables with respect to imputed PRA2-LFS segment}
\usage{
tabvisvar(varEsp, varEspText, vary, index, data, wvar = NULL, table = TRUE,
  plot = FALSE)
}
\arguments{
\item{varEsp}{character vector containing specific variables to tabulate}

\item{varEspText}{character vector containing text to print for the specific variable}

\item{vary}{character giving the name of the imputed LFS segment variable}

\item{index}{index of specific variable position in varesp}

\item{data}{data-frame containing the fused file}

\item{wvar}{weight variable, if weighting is desired}

\item{table}{logical value. TRUE if the table is desired, FALSE otherwise. Defaults to TRUE.}

\item{plot}{logical value. TRUE if a plot is desired, FALSE otherwise. Defaults to FALSE.}
}
\value{
list: table, plots
}
\description{
Tabulates a specific variable coming from Living Conditions Survey (ECV) with respect to the imputed segment from the Labour Force Survey (PRA)
}
\details{
Dependencies: ggplot2, gridExtra
}
\author{
Ines Garmendia
}

