% Generated by roxygen2 (4.0.2): do not edit by hand
\name{uncert2vars}
\alias{uncert2vars}
\title{Variable selection by uncertainty evaluation}
\usage{
uncert2vars(var_x, var_y, data_A, data_B, base = NULL, weights_A = NULL,
  weights_B = NULL, matchvars)
}
\arguments{
\item{var_x}{name of (specific) variable in the first file}

\item{var_y}{name of (specific) variable in the second file}

\item{data_A}{data-frame for the first file}

\item{data_B}{data-frame for the second file}

\item{base}{file from which to estimate marginal distributions for common variables: choose between values in data_A or data_B}

\item{weights_A}{name of weights variable in data1 (optional)}

\item{weights_B}{name of weights variable in data2 (optional)}

\item{matchvars}{list of common variables present in both files, candidates for matching}
}
\value{
list: name of varx, name of vary, best overall uncertainty, variable list for best ov.uncertainty
}
\description{
\code{uncert2vars} evaluates the uncertainty in the contingency table estimation between
two specific variables in separate files, given a list of common variables.
The function and selects the 'best' combination of common variables from the list
in terms of uncertainty reduction.
}
\details{
Uncertainty evaluation is based on Frechet bounds computation as implemented in
StatMatch \code{Fbwidths.by.x}
}
\seealso{
Other "Select matching variables": \code{\link{compare_matchvars}};
  \code{\link{compare_var_strata}};
  \code{\link{compare_var}}; \code{\link{plot2cat}};
  \code{\link{predictvalue_var}};
  \code{\link{predictvalue}}; \code{\link{similarity2cat}};
  \code{\link{tabulate2cat}}
}

