% Generated by roxygen2 (4.0.1): do not edit by hand
\name{uncertvarxvary}
\alias{uncertvarxvary}
\title{Variable selection by evaluating uncertainty in contingency table estimation}
\usage{
uncertvarxvary(varx, vary, data1, data2, basedata = NULL, varw1 = NULL,
  varw2 = NULL, varlist, ...)
}
\arguments{
\item{varx}{name of specific variable in the first file}

\item{vary}{name of specific variable in the second file}

\item{data1}{data-frame for the first file}

\item{data2}{data-frame for the second file}

\item{basedata}{data from which to estimate marginal distributions for commomn variables, choose between values given in data1 or data2}

\item{varw1}{name of weights variable in data1 (optional)}

\item{varw2}{name of weights variable in data2 (optional)}

\item{varlist}{list of common variables present in both files, candidates for matching}
}
\value{
list: name of varx, name of vary, best overall uncertainty, variable list for best ov.uncertainty
}
\description{
Computes uncertainty based on Frechet bounds for the contingency table for two specific variables (each in a file) given a list of common variables between the files, and selects the best combination of common variables from the list.
}
\details{
Dependencies: StatMatch
}
\seealso{
Other "Select variables": \code{\link{compareMultivar}};
  \code{\link{plot2Cat}}; \code{\link{plotTable}};
  \code{\link{predictvalue}}
}

