% Generated by roxygen2 (4.0.2): do not edit by hand
\name{update}
\alias{update}
\title{Update definition of a \code{filetomatch} object}
\usage{
update(x = "filetomatch", ...)
}
\arguments{
\item{x}{a \code{\link{filetomatch}} object}

\item{data}{new data frame}

\item{matchvars}{names of new matching variables}

\item{specvars}{names of new specific variables}

\item{stratavars}{names of new strata variables}

\item{weights}{names of weights variables}

\item{role}{new role: receptor, donor or symmetric}
}
\description{
\code{update} creates a new \code{\link{filetomatch}} object with updated values.

The user may want to change any of these values (or any combination) in the definition of a previously created
\code{filetomatch} object: \code{data}, \code{matchvars}, \code{specvars}, \code{weights}
or \code{role}.
}
\details{
Note that, in particular, the role of a \code{receptor} file can be changed to \code{donor}
or \code{symmetric}, or viceversa by updating the \code{role} value.

Note that the new object has to be stored (i.e. must be given a name).
}
\seealso{
Other "Utilities": \code{\link{include}};
  \code{\link{remove}}; \code{\link{select_observations}};
  \code{\link{select_strata}}
}

