% Generated by roxygen2 (4.0.2): do not edit by hand
\name{compare_matchvars}
\alias{compare_matchvars}
\title{Generic: "compare_matchvars"
applies to pairs of 'filetomatch' objects}
\usage{
compare_matchvars(x = "filetomatch", y = "filetomatch", ...)
}
\description{
Generic: "compare_matchvars"
applies to pairs of 'filetomatch' objects
}

