% Generated by roxygen2 (4.0.2): do not edit by hand
\name{compare_var_strata}
\alias{compare_var_strata}
\title{Generic: "compare_var_strata"
applies to pairs of 'filetomatch' objects}
\usage{
compare_var_strata(x = "filetomatch", y = "filetomatch", ...)
}
\description{
Generic: "compare_var_strata"
applies to pairs of 'filetomatch' objects
}

