% Generated by roxygen2 (4.0.2): do not edit by hand
\name{concatenate}
\alias{concatenate}
\title{Generic: "concatenate"
applies to pairs of 'filetomatch' objects}
\usage{
concatenate(x = "filetomatch", y = "filetomatch", ...)
}
\description{
Generic: "concatenate"
applies to pairs of 'filetomatch' objects
}

