% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{ecv}
\alias{ecv}
\title{Living Conditions Survey data frame}
\format{description A textual description of the format of the object.}
\source{
microdata for the Living Conditions Survey year 2009, Eustat, the Statistical Office of the Basque Government
}
\description{
Microdata file with the Living Conditions Survey 2009 questionnaire as provided by Eustat, the Statistical Office of the Basque Government
}
\details{
Microdata files
}
\author{
Ines Garmendia, 2014-07-11
}

