% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{filetomatch-class}
\alias{filetomatch-class}
\title{Microdata file to be matched: data plus information related to matching}
\description{
Microdata file to be matched: data plus information related to matching
}
\seealso{
Other "Matching definition step": \code{\link{fusedfile-class}};
  \code{\link{genericmatch-class}}
}

