% Generated by roxygen2 (4.0.2): do not edit by hand
\name{match.hotdeck}
\alias{match.hotdeck}
\title{Generic: "match"
 applies to pairs of 'filetomatch' objects}
\usage{
match.hotdeck(x = "filetomatch", y = "filetomatch", ...)
}
\description{
Generic: "match"
 applies to pairs of 'filetomatch' objects
}

