% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plot2Cat}
\alias{plot2Cat}
\title{Grouped bars to compare two categorical variables}
\usage{
plot2Cat(table1, table2, type = "rel")
}
\arguments{
\item{table1}{distribution of the variable in the 1st file, result of as.data.frame(xtabs()). Counts (type='abs') or freqs (type='rel')}

\item{table2}{distribution of the variable in the 2nd file, result of as.data.frame(xtabs()). Counts (type='abs') or freqs (type='rel')}

\item{type}{type of values associated with the levels of the factor: absolute (abs) or relative (rel)}
}
\value{
a graphical object (barplot)
}
\description{
Internal function that computes grouped bars for comparting a factor variable between two files,
to be used inside other functions in micromatch.
}
\details{
Details here if necessary.
}
\author{
"Ines Garmendia <ines.garmendia@gmail.com>"
}
\seealso{
Other "Select variables": \code{\link{compareMultivar}};
  \code{\link{plotTable}}; \code{\link{predictvalue}};
  \code{\link{uncertvarxvary}}
}
\keyword{graphics}

