% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plot2cat}
\alias{plot2cat}
\title{Compare a single categorical variable across two data frames: Plot}
\usage{
plot2cat(data_A, data_B, var_A, var_B, weights_A = NULL, weights_B = NULL,
  cell_values = "abs")
}
\description{
Barplots for a single categorical variable in two distinct, independent data frames
}
\seealso{
Other 'Select matching variables': \code{\link{compare_matchvars,filetomatch,filetomatch-method}};
  \code{\link{compare_var,filetomatch,filetomatch-method}};
  \code{\link{compare_var_strata,filetomatch,filetomatch-method}};
  \code{\link{predictvalue,filetomatch-method}};
  \code{\link{similarity2cat}}
}

