% Generated by roxygen2 (4.0.2): do not edit by hand
\name{predictvalue}
\alias{predictvalue}
\title{Assess predictive value of common variables w.r.t specific variables in a \code{filetomatch} object}
\usage{
predictvalue(x = "filetomatch", ...)
}
\arguments{
\item{x}{\code{filetomatch} object}
}
\value{
\code{predictvalue} function returns a list with measures for
each combination of variables in \code{matchvars} and \code{specvars}.

Specifically, for each combination it returns a list with four values:
\itemize{
     \item \code{V}: a vector with the estimated Cramer's V for each couple response-predictor.
     \item \code{lambda}: a vector with the values of Goodman-Kruscal lambda(R|C) for each couple response-predictor.
     \item \code{tau}: a vector with the values of Goodman-Kruscal tau(R|C) for each couple response-predictor.
     \item \code{U}: a vector whit the values of Theil's uncertainty coefficient U(R|C) for each couple response-predictor.
}
}
\description{
\code{predictvalue} evaluates the predictive value for each
\code{matchvars} variable with respect to the \code{specvars} in a
\code{\link{filetomatch}} object.
}
\seealso{
Other "Select matching variables": \code{\link{compare_matchvars}};
  \code{\link{compare_var_strata}};
  \code{\link{compare_var}}; \code{\link{plot2cat}};
  \code{\link{predictvalue_var}};
  \code{\link{similarity2cat}}; \code{\link{tabulate2cat}};
  \code{\link{uncert2vars}}
}

