% Generated by roxygen2 (4.0.2): do not edit by hand
\name{predictvalue_var}
\alias{predictvalue_var}
\title{Assess predictive value of a variable w.r.t. another variable in a data frame}
\usage{
predictvalue_var(data, vars_x, var_y, weights = NULL)
}
\arguments{
\item{data}{name of data frame containing both x and y}

\item{vars_x}{name of common variable(s)}

\item{var_y}{name of a specific variable}

\item{weights}{weight vector name (optional)}
}
\value{
A list with four components:
\itemize{
     \item \code{V} A vector with the estimated Cramer's V for each couple response-predictor.
     \item \code{lambda} A vector with the values of Goodman-Kruscal lambda(R|C) for each couple response-predictor.
     \item \code{tau} A vector with the values of Goodman-Kruscal tau(R|C) for each couple response-predictor.
     \item \code{U} A vector whit the values of Theil's uncertainty coefficient U(R|C) for each couple response-predictor.
}
}
\description{
Evaluates predictive value of a common variable \code{x}
with respect to a specific variable \code{y}
}
\details{
This function relies on function \code{pw.assoc()} from
\code{StatMatch} package.
}
\seealso{
Other "Select matching variables": \code{\link{compare_matchvars}};
  \code{\link{compare_var_strata}};
  \code{\link{compare_var}}; \code{\link{plot2cat}};
  \code{\link{predictvalue}}; \code{\link{similarity2cat}};
  \code{\link{tabulate2cat}}; \code{\link{uncert2vars}}
}

