% Generated by roxygen2 (4.0.2): do not edit by hand
\name{select_observations}
\alias{select_observations}
\title{Generic: "select_observations"
applies to a 'filetomatch' object}
\usage{
select_observations(x = "filetomatch", ...)
}
\description{
Generic: "select_observations"
applies to a 'filetomatch' object
}

